Tired of waiting for couch
I have a program running on my desktop that pulls the ftp file, saves
it in a db, finds a route based on older records, and then only gives
me one json object for each bus. This leaves little processing to the
client side. Changed the client to handle this different data stream.
Also played with the UI a bit.